create context service for markdown.extension.editor.cursor.inMarkdownList context
